+++ index.d.ts
@@ -97,6 +97,10 @@ interface Document {
 	};
 }
 
+interface RenderContentFn {
+	(props: Dictionary): string;
+}
+
 interface RenderParams {
 	/**
 	 * Component attrs
@@ -104,7 +108,7 @@ interface RenderParams {
 	attrs: Dictionary;
 
 	/** @see [[RenderContent]] */
-	content?: Dictionary<RenderContent | string>;
+	content?: Dictionary<RenderContent | RenderContentFn | string>;
 }
 
 /**
@@ -149,7 +153,7 @@ interface RenderContent {
 	attrs: Dictionary;
 
 	/** @see [[RenderContent]] */
-	content?: Dictionary<RenderContent | string>;
+	content?: Dictionary<RenderContent | RenderContentFn | string>;
 }
 
 // eslint-disable-next-line no-var,vars-on-top