+++ src/base/b-router/test/index.js
@@ -8,9 +8,19 @@
 
 /* eslint-disable max-lines-per-function */
 
+/**
+ * @typedef {import('playwright').Page} Page
+ */
+
 const
 	h = include('tests/helpers');
 
+/**
+ * Starts a test
+ *
+ * @param {Page} page
+ * @returns {!Promise<void>}
+ */
 module.exports = async (page) => {
 	const
 		// eslint-disable-next-line import/no-nodejs-modules