+++ src/icons/b-icon/b-icon.ts
@@ -13,7 +13,6 @@
 
 import iIcon from 'traits/i-icon/i-icon';
 import iSize from 'traits/i-size/i-size';
-
 import iBlock, { component, prop, ModsDecl } from 'super/i-block/i-block';
 
 export * from 'super/i-block/i-block';
@@ -34,18 +33,6 @@ export default class bIcon extends iBlock implements iIcon, iSize {
 	 */
 	@prop(String)
 	readonly prfx: string = '';
-
-	/**
-	 * Tooltip text
-	 */
-	@prop({type: String, required: false})
-	readonly hint?: string;
-
-	/**
-	 * Tooltip position
-	 */
-	@prop(String)
-	readonly hintPos: string = 'bottom';
 
 	/** @override */
 	static readonly mods: ModsDecl = {