+++ src/pages/p-v4-components-demo/index.js
@@ -11,4 +11,11 @@ const
 
 package('p-v4-components-demo')
 	.extends('i-static-page')
-	.dependencies('b-v4-component-demo', 'b-virtual-scroll', 'b-button', 'b-router', components);
+	.dependencies(
+		'b-v4-component-demo',
+		'b-virtual-scroll',
+		'b-button',
+		'b-router',
+		'b-list',
+		components
+	);