+++ src/super/i-static-page/i-static-page.ts
@@ -279,7 +279,7 @@ export default abstract class iStaticPage extends iPage {
 	}
 
 	/**
-	 * Synchronization for .localeStore field
+	 * Synchronization of the "localeStore" field
 	 * @param locale
 	 */
 	@watch(['localeStore', 'globalEmitter:i18n.setLocale'])
@@ -293,7 +293,7 @@ export default abstract class iStaticPage extends iPage {
 	}
 
 	/**
-	 * Synchronization for .isAuth field
+	 * Synchronization of the "isAuth" field
 	 * @param [e]
 	 */
 	@watch('globalEmitter:session.*')
@@ -302,7 +302,7 @@ export default abstract class iStaticPage extends iPage {
 	}
 
 	/**
-	 * Synchronization for .isOnline field
+	 * Synchronization of the "isOnline" field
 	 * @param e
 	 */
 	@watch('globalEmitter:net.status')