+++ src/core/prelude/test-env/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.90 (2020-10-22)
+
+#### :rocket: New Feature
+
+* Now `content` can be provided as a function `renderComponents`
+
 ## v3.0.0-rc.85 (2020-10-09)
 
 #### :house: Internal