+++ src/super/i-block/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.90 (2020-10-22)
+
+#### :boom: Breaking Change
+
+* Now all extra classes that were added by using `appendToRootClasses` added to the start of the declaration
+
 ## v3.0.0-rc.88 (2020-10-13)
 
 #### :bug: Bug Fix